WIP: uint32_t test

+++ src/drivers/pc80/vga/vga.c
@@ -350,7 +350,7 @@ vga_textmode_init(void)
 	vga_write_text(VGA_TEXT_LEFT, VGA_LINES - 1, test_str);
 	vga_write_text(VGA_TEXT_RIGHT, VGA_LINES - 1, test_str);
 
-	uint16_t *p = (uint16_t *)VGA_FB;
+	uint32_t *p = (uint32_t *)VGA_FB;
 	for (uint16_t i = 0; i < UINT16_MAX; i++) {
 		p[i % (VGA_COLUMNS * VGA_LINES)] = i;
 		if (i % (VGA_COLUMNS * VGA_LINES) == 0)
